move `OriginalSymbolsMatch` to `DependentTypeFinder` and remove `ref` parameter
