+++ src/Workspaces/Core/Portable/FindSymbols/FindReferences/Finders/AbstractReferenceFinder.cs
@@ -217,8 +217,6 @@ namespace Microsoft.CodeAnalysis.FindSymbols.Finders
         protected static Func<SyntaxNode, SemanticModel, ValueTuple<bool, CandidateReason>> GetStandardSymbolsNodeMatchFunction(
             ISymbol searchSymbol, Solution solution, CancellationToken cancellationToken)
         {
-            Compilation searchSymbolCompilation = null;
-
             Func<SyntaxNode, SemanticModel, ValueTuple<bool, CandidateReason>> symbolsMatch =
                 (node, model) =>
                 {
@@ -227,11 +225,11 @@ namespace Microsoft.CodeAnalysis.FindSymbols.Finders
                     var symbolToMatch = symbolInfoToMatch.Symbol;
                     var symbolToMatchCompilation = model.Compilation;
 
-                    if (OriginalSymbolsMatch(searchSymbol, symbolInfoToMatch.Symbol, solution, ref searchSymbolCompilation, symbolToMatchCompilation, cancellationToken))
+                    if (DependentTypeFinder.OriginalSymbolsMatch(searchSymbol, symbolInfoToMatch.Symbol, solution, null, symbolToMatchCompilation, cancellationToken))
                     {
                         return ValueTuple.Create(true, CandidateReason.None);
                     }
-                    else if (symbolInfoToMatch.CandidateSymbols.Any(s => OriginalSymbolsMatch(searchSymbol, s, solution, ref searchSymbolCompilation, symbolToMatchCompilation, cancellationToken)))
+                    else if (symbolInfoToMatch.CandidateSymbols.Any(s => DependentTypeFinder.OriginalSymbolsMatch(searchSymbol, s, solution, null, symbolToMatchCompilation, cancellationToken)))
                     {
                         return ValueTuple.Create(true, symbolInfoToMatch.CandidateReason);
                     }
@@ -242,267 +240,6 @@ namespace Microsoft.CodeAnalysis.FindSymbols.Finders
                 };
 
             return symbolsMatch;
-        }
-
-        internal static bool OriginalSymbolsMatch(
-            ISymbol searchSymbol,
-            ISymbol symbolToMatch,
-            Solution solution,
-            CancellationToken cancellationToken)
-        {
-            if (ReferenceEquals(searchSymbol, symbolToMatch))
-            {
-                return true;
-            }
-
-            if (searchSymbol == null || symbolToMatch == null)
-            {
-                return false;
-            }
-
-            Compilation searchSymbolCompilation = null;
-            if (!TryGetCompilation(searchSymbol, solution, ref searchSymbolCompilation, cancellationToken))
-            {
-                return false;
-            }
-
-            Compilation symbolToMatchCompilation = null;
-            if (!TryGetCompilation(symbolToMatch, solution, ref symbolToMatchCompilation, cancellationToken))
-            {
-                return false;
-            }
-
-            return OriginalSymbolsMatch(searchSymbol, symbolToMatch, solution, ref searchSymbolCompilation, symbolToMatchCompilation, cancellationToken);
-        }
-
-        private static bool OriginalSymbolsMatch(
-            ISymbol searchSymbol,
-            ISymbol symbolToMatch,
-            Solution solution,
-            ref Compilation searchSymbolCompilation,
-            Compilation symbolToMatchCompilation,
-            CancellationToken cancellationToken)
-        {
-            if (symbolToMatch == null)
-            {
-                return false;
-            }
-
-            if (OriginalSymbolsMatchCore(searchSymbol, symbolToMatch, solution, ref searchSymbolCompilation, symbolToMatchCompilation, cancellationToken))
-            {
-                return true;
-            }
-
-            if (searchSymbol.Kind == SymbolKind.Namespace && symbolToMatch.Kind == SymbolKind.Namespace)
-            {
-                // if one of them is a merged namespace symbol and other one is its constituent namespace symbol, they are equivalent.
-                var namespace1 = (INamespaceSymbol)searchSymbol;
-                var namespace2 = (INamespaceSymbol)symbolToMatch;
-                var namespace1Count = namespace1.ConstituentNamespaces.Length;
-                var namespace2Count = namespace2.ConstituentNamespaces.Length;
-                if (namespace1Count != namespace2Count)
-                {
-                    if ((namespace1Count > 1 &&
-                         namespace1.ConstituentNamespaces.Any(n => NamespaceSymbolsMatch(n, namespace2, solution, cancellationToken))) ||
-                        (namespace2Count > 1 &&
-                         namespace2.ConstituentNamespaces.Any(n2 => NamespaceSymbolsMatch(namespace1, n2, solution, cancellationToken))))
-                    {
-                        return true;
-                    }
-                }
-            }
-
-            if (searchSymbol.Kind == SymbolKind.NamedType && symbolToMatch.IsConstructor())
-            {
-                return OriginalSymbolsMatch(searchSymbol, symbolToMatch.ContainingType, solution, ref searchSymbolCompilation, symbolToMatchCompilation, cancellationToken);
-            }
-
-            return false;
-        }
-
-        private static bool NamespaceSymbolsMatch(
-            INamespaceSymbol namespace1,
-            INamespaceSymbol namespace2,
-            Solution solution,
-            CancellationToken cancellationToken)
-        {
-            Compilation definingCompilation = null;
-            return OriginalSymbolsMatch(namespace1, namespace2, solution, ref definingCompilation, null, cancellationToken);
-        }
-
-        private static bool OriginalSymbolsMatchCore(
-            ISymbol searchSymbol,
-            ISymbol symbolToMatch,
-            Solution solution,
-            ref Compilation searchSymbolCompilation,
-            Compilation symbolToMatchCompilation,
-            CancellationToken cancellationToken)
-        {
-            if (searchSymbol == null || symbolToMatch == null)
-            {
-                return false;
-            }
-
-            searchSymbol = searchSymbol.GetOriginalUnreducedDefinition();
-            symbolToMatch = symbolToMatch.GetOriginalUnreducedDefinition();
-
-            // We compare the given searchSymbol and symbolToMatch for equivalence using SymbolEquivalenceComparer
-            // as follows:
-            //  1)  We compare the given symbols using the SymbolEquivalenceComparer.IgnoreAssembliesInstance,
-            //      which ignores the containing assemblies for named types equivalence checks. This is required
-            //      to handle equivalent named types which are forwarded to completely different assemblies.
-            //  2)  If the symbols are NOT equivalent ignoring assemblies, then they cannot be equivalent.
-            //  3)  Otherwise, if the symbols ARE equivalent ignoring assemblies, they may or may not be equivalent
-            //      if containing assemblies are NOT ignored. We need to perform additional checks to ensure they
-            //      are indeed equivalent:
-            //
-            //      (a) If IgnoreAssembliesInstance.Equals equivalence visitor encountered any pair of non-nested 
-            //          named types which were equivalent in all aspects, except that they resided in different 
-            //          assemblies, we need to ensure that all such pairs are indeed equivalent types. Such a pair
-            //          of named types is equivalent if and only if one of them is a type defined in either 
-            //          searchSymbolCompilation(C1) or symbolToMatchCompilation(C2), say defined in reference assembly
-            //          A (version v1) in compilation C1, and the other type is a forwarded type, such that it is 
-            //          forwarded from reference assembly A (version v2) to assembly B in compilation C2.
-            //      (b) Otherwise, if no such named type pairs were encountered, symbols ARE equivalent.
-
-            using (var equivalentTypesWithDifferingAssemblies = SharedPools.Default<Dictionary<INamedTypeSymbol, INamedTypeSymbol>>().GetPooledObject())
-            {
-                // 1) Compare searchSymbol and symbolToMatch using SymbolEquivalenceComparer.IgnoreAssembliesInstance
-                if (!SymbolEquivalenceComparer.IgnoreAssembliesInstance.Equals(searchSymbol, symbolToMatch, equivalentTypesWithDifferingAssemblies.Object))
-                {
-                    // 2) If the symbols are NOT equivalent ignoring assemblies, then they cannot be equivalent.
-                    return false;
-                }
-
-                // 3) If the symbols ARE equivalent ignoring assemblies, they may or may not be equivalent if containing assemblies are NOT ignored.
-                if (equivalentTypesWithDifferingAssemblies.Object.Count > 0)
-                {
-                    // Step 3a) Ensure that all pairs of named types in equivalentTypesWithDifferingAssemblies are indeed equivalent types.
-                    return VerifyForwardedTypes(equivalentTypesWithDifferingAssemblies.Object, searchSymbol, symbolToMatch,
-                        solution, ref searchSymbolCompilation, symbolToMatchCompilation, cancellationToken);
-                }
-
-                // 3b) If no such named type pairs were encountered, symbols ARE equivalent.
-                return true;
-            }
-        }
-
-        // Verifies that all pairs of named types in equivalentTypesWithDifferingAssemblies are equivalent forwarded types.
-        private static bool VerifyForwardedTypes(
-            Dictionary<INamedTypeSymbol, INamedTypeSymbol> equivalentTypesWithDifferingAssemblies,
-            ISymbol searchSymbol,
-            ISymbol symbolToMatch,
-            Solution solution,
-            ref Compilation searchSymbolCompilation,
-            Compilation symbolToMatchCompilation,
-            CancellationToken cancellationToken)
-        {
-            var verifiedKeys = new HashSet<INamedTypeSymbol>();
-            var count = equivalentTypesWithDifferingAssemblies.Count;
-            int verifiedCount = 0;
-
-            // First check forwarded types in searchSymbolCompilation.
-            if (TryGetCompilation(searchSymbol, solution, ref searchSymbolCompilation, cancellationToken))
-            {
-                verifiedCount = VerifyForwardedTypes(equivalentTypesWithDifferingAssemblies, searchSymbolCompilation, verifiedKeys, isSearchSymbolCompilation: true);
-                if (verifiedCount == count)
-                {
-                    // All equivalent types verified.
-                    return true;
-                }
-            }
-
-            // Now check forwarded types in symbolToMatchCompilation.
-            verifiedCount += VerifyForwardedTypes(equivalentTypesWithDifferingAssemblies, symbolToMatchCompilation, verifiedKeys, isSearchSymbolCompilation: false);
-            return verifiedCount == count;
-        }
-
-        private static int VerifyForwardedTypes(
-            Dictionary<INamedTypeSymbol, INamedTypeSymbol> equivalentTypesWithDifferingAssemblies,
-            Compilation compilation,
-            HashSet<INamedTypeSymbol> verifiedKeys,
-            bool isSearchSymbolCompilation)
-        {
-            Contract.ThrowIfNull(compilation);
-            Contract.ThrowIfNull(equivalentTypesWithDifferingAssemblies);
-            Contract.ThrowIfTrue(!equivalentTypesWithDifferingAssemblies.Any());
-
-            // Must contain equivalents named types residing in different assemblies.
-            Contract.ThrowIfFalse(equivalentTypesWithDifferingAssemblies.All(kvp => !SymbolEquivalenceComparer.Instance.Equals(kvp.Key.ContainingAssembly, kvp.Value.ContainingAssembly)));
-
-            // Must contain non-nested named types.
-            Contract.ThrowIfFalse(equivalentTypesWithDifferingAssemblies.All(kvp => kvp.Key.ContainingType == null));
-            Contract.ThrowIfFalse(equivalentTypesWithDifferingAssemblies.All(kvp => kvp.Value.ContainingType == null));
-
-            var referencedAssemblies = new MultiDictionary<string, IAssemblySymbol>();
-            foreach (var assembly in compilation.GetReferencedAssemblySymbols())
-            {
-                referencedAssemblies.Add(assembly.Name, assembly);
-            }
-
-            int verifiedCount = 0;
-            foreach (var kvp in equivalentTypesWithDifferingAssemblies)
-            {
-                if (!verifiedKeys.Contains(kvp.Key))
-                {
-                    INamedTypeSymbol originalType, expectedForwardedType;
-                    if (isSearchSymbolCompilation)
-                    {
-                        originalType = kvp.Value.OriginalDefinition;
-                        expectedForwardedType = kvp.Key.OriginalDefinition;
-                    }
-                    else
-                    {
-                        originalType = kvp.Key.OriginalDefinition;
-                        expectedForwardedType = kvp.Value.OriginalDefinition;
-                    }
-
-                    foreach (var referencedAssembly in referencedAssemblies[originalType.ContainingAssembly.Name])
-                    {
-                        var fullyQualifiedTypeName = originalType.MetadataName;
-                        if (originalType.ContainingNamespace != null)
-                        {
-                            fullyQualifiedTypeName = originalType.ContainingNamespace.ToDisplayString(SymbolDisplayFormats.SignatureFormat) +
-                                "." + fullyQualifiedTypeName;
-                        }
-
-                        // Resolve forwarded type and verify that the types from different assembly are indeed equivalent.
-                        var forwardedType = referencedAssembly.ResolveForwardedType(fullyQualifiedTypeName);
-                        if (forwardedType == expectedForwardedType)
-                        {
-                            verifiedKeys.Add(kvp.Key);
-                            verifiedCount++;
-                        }
-                    }
-                }
-            }
-
-            return verifiedCount;
-        }
-
-        private static bool TryGetCompilation(
-            ISymbol symbol,
-            Solution solution,
-            ref Compilation definingCompilation,
-            CancellationToken cancellationToken)
-        {
-            if (definingCompilation == null)
-            {
-                var definitionProject = solution.GetProject(symbol.ContainingAssembly, cancellationToken);
-                if (definitionProject == null)
-                {
-                    return false;
-                }
-
-                // compilation from definition project must already exist.
-                if (!definitionProject.TryGetCompilation(out definingCompilation))
-                {
-                    Contract.Requires(false, "How can compilation not exist?");
-                    return false;
-                }
-            }
-
-            return true;
         }
 
         protected Task<IEnumerable<ReferenceLocation>> FindReferencesInTokensAsync(
